Extract utxoCommitment and utxoNullifier to a higher level
