+++ integration-tests/src/test/no-proving/DustLocalState.test.ts
@@ -34,7 +34,9 @@ import {
   type UtxoOutput,
   type UtxoSpend,
   WellFormedStrictness,
-  updatedValue
+  updatedValue,
+  utxoCommitment,
+  utxoNullifier
 } from '@midnight-ntwrk/ledger';
 import { expect } from 'vitest';
 import { ProofMarker, SignatureMarker } from '@/test/utils/Markers';
@@ -1029,7 +1031,7 @@ describe('Ledger API - DustLocalState', () => {
     const { secretKey } = state.dustKey;
 
     const qdo = localState.utxos[0];
-    const nullifier = localState.utxoNullifier(qdo, secretKey);
+    const nullifier = utxoNullifier(qdo, secretKey);
 
     // Create a new local state and manually add the UTXO
     const newLocalState = new DustLocalState(initialParameters);
@@ -1062,7 +1064,7 @@ describe('Ledger API - DustLocalState', () => {
 
     const qdo = localState.utxos[0];
     const genInfo = state.dust.generationInfo(qdo)!;
-    const nullifier = localState.utxoNullifier(qdo, secretKey);
+    const nullifier = utxoNullifier(qdo, secretKey);
 
     // Create a new local state and add the UTXO with a pending timestamp
     const pendingUntil = new Date(now.getTime() + 1000);
@@ -1095,7 +1097,7 @@ describe('Ledger API - DustLocalState', () => {
     const { secretKey } = state.dustKey;
 
     const qdo = localState.utxos[0];
-    const nullifier = localState.utxoNullifier(qdo, secretKey);
+    const nullifier = utxoNullifier(qdo, secretKey);
 
     // Add the UTXO to a new local state
     const newLocalState = new DustLocalState(initialParameters);
@@ -1127,7 +1129,7 @@ describe('Ledger API - DustLocalState', () => {
     const localState = state.dust;
 
     const qdo = localState.utxos[0];
-    const commitment = localState.utxoCommitment(qdo);
+    const commitment = utxoCommitment(qdo);
 
     expect(commitment).toBeDefined();
     expect(typeof commitment).toBe('bigint');
@@ -1146,7 +1148,7 @@ describe('Ledger API - DustLocalState', () => {
     const { secretKey } = state.dustKey;
 
     const qdo = localState.utxos[0];
-    const nullifier = localState.utxoNullifier(qdo, secretKey);
+    const nullifier = utxoNullifier(qdo, secretKey);
 
     expect(nullifier).toBeDefined();
     expect(typeof nullifier).toBe('bigint');
@@ -1182,8 +1184,8 @@ describe('Ledger API - DustLocalState', () => {
     expect(newUtxo.backingNight).toEqual(qdo.backingNight);
 
     // The new UTXO should have a different commitment
-    const originalCommitment = localState.utxoCommitment(qdo);
-    const newCommitment = localState.utxoCommitment(newUtxo);
+    const originalCommitment = utxoCommitment(qdo);
+    const newCommitment = utxoCommitment(newUtxo);
     expect(newCommitment).not.toEqual(originalCommitment);
   });
 });
